% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_jar_path}
\alias{get_jar_path}
\title{Find the path to the LanguageTool jar file}
\usage{
get_jar_path()
}
\value{
A string indicating the path to the jar file
}
\description{
Find the path to the LanguageTool jar file in the local installation of the package
}
