% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_lang_list}
\alias{get_lang_list}
\title{Provide the avaiblable languages}
\usage{
get_lang_list()
}
\value{
A character vector
}
\description{
Provide the avaiblable languages if the local file containing them exists
}
\examples{
\donttest{
get_lang_list()
}
}
\seealso{
[install_lang_list()]
}
