% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{install_lang_list}
\alias{install_lang_list}
\title{Create a local file with the available languages}
\usage{
install_lang_list(port = 8805)
}
\arguments{
\item{port}{The port of the server}
}
\value{
Nothing
}
\description{
Request a running server to save the available languages in a local file
}
\examples{
\donttest{
starting_server(8805)
install_lang_list(8805)
}
}
\seealso{
[get_lang_list()]
}
