% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{install_LanguageTools}
\alias{install_LanguageTools}
\title{Download the LanguageTool jar file}
\usage{
install_LanguageTools()
}
\value{
TRUE if the installation worked, FALSE otherwise
}
\description{
Download and unzip a jar file of LanguageTool (https://languagetool.org/download/LanguageTool-stable.zip)
}
\examples{
\donttest{
install_LanguageTools()
}
}
