% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{is_LanguageTools_installed}
\alias{is_LanguageTools_installed}
\title{Check is LanguageTools is installed}
\usage{
is_LanguageTools_installed()
}
\value{
A boolean
}
\description{
Check if LanguageTools has been installed locally
}
\examples{
\donttest{
is_LanguageTools_installed()
}
}
