% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\docType{data}
\name{Language_env}
\alias{Language_env}
\title{langCorrector general environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
Language_env
}
\description{
An environment used by langCorrector to store data, functions and values
}
\keyword{datasets}
