% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{rmd_corrector}
\alias{rmd_corrector}
\title{Rmarkdown corrector Shiny app}
\usage{
rmd_corrector(rmdfile, language, port = 8805)
}
\arguments{
\item{rmdfile}{Astring, the path to the rmd file to edit}

\item{language}{A string, the language to use by LanguageTools}

\item{port}{The port to use for the LanguageTools server}
}
\value{
Nothing
}
\description{
Start a Shiny app to correct and edit Rmarkdown files
}
\examples{
\donttest{
rmd_corrector("path/to/myfile.rmd", "EN", 8805)
}
}
