% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{running_server_id}
\alias{running_server_id}
\title{Check if a current process could be the LanguageTools server}
\usage{
running_server_id()
}
\value{
A ps object representing the server or NULL
}
\description{
Check if a current process could be the LanguageTools server
}
