% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{send_request}
\alias{send_request}
\title{Send a request to a running LanguageTools server}
\usage{
send_request(port = 8805, text, language)
}
\arguments{
\item{port}{The port of the server}

\item{text}{The string to check}

\item{language}{A string indicating which language to use}
}
\value{
A list with the corrections from LanguageTools
}
\description{
Send a text for correction to a running LanguageTools server
}
\examples{
\donttest{
starting_server(8805)
corrections <- send_request(8805, "Hello there, it iss a obvious mistaken.", language = "EN")
}
}
\seealso{
[install_lang_list()] and [get_lang_list()]
}
